requirements.txt
alembic==0.9.1
setuptools==41.2.0
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.1.2
MarkupSafe==1.0
packaging==16.8
psycopg2==2.8.5
pyparsing==2.2.0
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
opentelemetry-ext-flask==0.8b0
opentelemetry-ext-jaeger==0.8b0
opentelemetry-sdk==0.8b0
opentelemetry-api==0.8b0